# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeApp API — skills, tasks, notes, translation, AI assistant
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  pending: 'pending',
  completed: 'completed',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  category?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  priority?: TaskInputPriority;
  dueDate?: string;
  category?: string;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  pending: 'pending',
  completed: 'completed',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  dueDate?: string;
  category?: string;
}

export interface TaskStats {
  total: number;
  completed: number;
  pending: number;
  highPriority: number;
  completionRate: number;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  tag?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  content: string;
  tag?: string;
}

export interface NoteUpdate {
  title?: string;
  content?: string;
  tag?: string;
}

export interface Skill {
  id: number;
  name: string;
  description: string;
  category: string;
  icon: string;
  totalLessons: number;
  completedLessons: number;
}

export interface Lesson {
  id: number;
  skillId: number;
  title: string;
  content: string;
  order: number;
  isCompleted: boolean;
  /** @nullable */
  duration?: string | null;
}

export interface SkillDetail {
  id: number;
  name: string;
  description: string;
  category: string;
  icon: string;
  totalLessons: number;
  completedLessons: number;
  lessons: Lesson[];
}

export interface LessonProgressInput {
  lessonId: number;
  isCompleted: boolean;
}

export interface LessonProgress {
  lessonId: number;
  skillId: number;
  isCompleted: boolean;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface ChatInput {
  /** @minLength 1 */
  message: string;
}

export interface TranslationInput {
  /** @minLength 1 */
  text: string;
  targetLanguage: string;
  sourceLanguage?: string;
}

export interface TranslationResult {
  originalText: string;
  translatedText: string;
  sourceLanguage: string;
  targetLanguage: string;
}

export interface TranslationRecord {
  id: number;
  originalText: string;
  translatedText: string;
  sourceLanguage: string;
  targetLanguage: string;
  createdAt: string;
}

export type DashboardSummarySkills = {
  totalSkills: number;
  totalCompleted: number;
  totalInProgress: number;
};

export type DashboardSummaryNotes = {
  total: number;
  recentCount: number;
};

export type DashboardSummaryRecentActivityItem = {
  type: string;
  description: string;
  timestamp: string;
};

export interface DashboardSummary {
  tasks: TaskStats;
  skills: DashboardSummarySkills;
  notes: DashboardSummaryNotes;
  recentActivity: DashboardSummaryRecentActivityItem[];
}

export type ListTasksParams = {
status?: ListTasksStatus;
priority?: ListTasksPriority;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  pending: 'pending',
  completed: 'completed',
  all: 'all',
} as const;

export type ListTasksPriority = typeof ListTasksPriority[keyof typeof ListTasksPriority];


export const ListTasksPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;