# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LifeApp API — skills, tasks, notes, translation, AI assistant
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Daily task management
  - name: notes
    description: Personal notes
  - name: skills
    description: Everyday skills learning
  - name: chat
    description: AI assistant chat
  - name: translate
    description: Language translation
  - name: dashboard
    description: Dashboard summary

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, all]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /tasks/stats:
    get:
      operationId: getTaskStats
      tags: [tasks]
      summary: Get task statistics
      responses:
        "200":
          description: Task stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStats"

  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  /notes/{id}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List all skill categories
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"

  /skills/{id}:
    get:
      operationId: getSkill
      tags: [skills]
      summary: Get a skill with its lessons
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Skill with lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /skills/{id}/progress:
    post:
      operationId: markLessonComplete
      tags: [skills]
      summary: Mark a lesson as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonProgressInput"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonProgress"

  /chat:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: Get chat history
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  /chat/clear:
    delete:
      operationId: clearChat
      tags: [chat]
      summary: Clear all chat history
      responses:
        "204":
          description: Cleared

  /translate:
    post:
      operationId: translateText
      tags: [translate]
      summary: Translate text between languages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslationInput"
      responses:
        "200":
          description: Translation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslationResult"

  /translate/history:
    get:
      operationId: listTranslationHistory
      tags: [translate]
      summary: Get translation history
      responses:
        "200":
          description: Translation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TranslationRecord"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]

    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]

    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
        category:
          type: string

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [pending, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
        category:
          type: string

    TaskStats:
      type: object
      required: [total, completed, pending, highPriority, completionRate]
      properties:
        total:
          type: integer
        completed:
          type: integer
        pending:
          type: integer
        highPriority:
          type: integer
        completionRate:
          type: number

    Note:
      type: object
      required: [id, title, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        tag:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    NoteInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        tag:
          type: string

    NoteUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tag:
          type: string

    Skill:
      type: object
      required: [id, name, description, category, icon, totalLessons, completedLessons]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: string
        totalLessons:
          type: integer
        completedLessons:
          type: integer

    Lesson:
      type: object
      required: [id, skillId, title, content, order, isCompleted]
      properties:
        id:
          type: integer
        skillId:
          type: integer
        title:
          type: string
        content:
          type: string
        order:
          type: integer
        isCompleted:
          type: boolean
        duration:
          type: ["string", "null"]

    SkillDetail:
      type: object
      required: [id, name, description, category, icon, totalLessons, completedLessons, lessons]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: string
        totalLessons:
          type: integer
        completedLessons:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"

    LessonProgressInput:
      type: object
      required: [lessonId, isCompleted]
      properties:
        lessonId:
          type: integer
        isCompleted:
          type: boolean

    LessonProgress:
      type: object
      required: [lessonId, skillId, isCompleted]
      properties:
        lessonId:
          type: integer
        skillId:
          type: integer
        isCompleted:
          type: boolean

    ChatMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string

    ChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1

    TranslationInput:
      type: object
      required: [text, targetLanguage]
      properties:
        text:
          type: string
          minLength: 1
        targetLanguage:
          type: string
        sourceLanguage:
          type: string

    TranslationResult:
      type: object
      required: [originalText, translatedText, sourceLanguage, targetLanguage]
      properties:
        originalText:
          type: string
        translatedText:
          type: string
        sourceLanguage:
          type: string
        targetLanguage:
          type: string

    TranslationRecord:
      type: object
      required: [id, originalText, translatedText, sourceLanguage, targetLanguage, createdAt]
      properties:
        id:
          type: integer
        originalText:
          type: string
        translatedText:
          type: string
        sourceLanguage:
          type: string
        targetLanguage:
          type: string
        createdAt:
          type: string

    DashboardSummary:
      type: object
      required: [tasks, skills, notes, recentActivity]
      properties:
        tasks:
          $ref: "#/components/schemas/TaskStats"
        skills:
          type: object
          required: [totalSkills, totalCompleted, totalInProgress]
          properties:
            totalSkills:
              type: integer
            totalCompleted:
              type: integer
            totalInProgress:
              type: integer
        notes:
          type: object
          required: [total, recentCount]
          properties:
            total:
              type: integer
            recentCount:
              type: integer
        recentActivity:
          type: array
          items:
            type: object
            required: [type, description, timestamp]
            properties:
              type:
                type: string
              description:
                type: string
              timestamp:
                type: string