# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeApp API — skills, tasks, notes, translation, AI assistant
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryNotes } from './dashboardSummaryNotes';
import type { DashboardSummaryRecentActivityItem } from './dashboardSummaryRecentActivityItem';
import type { DashboardSummarySkills } from './dashboardSummarySkills';
import type { TaskStats } from './taskStats';

export interface DashboardSummary {
  tasks: TaskStats;
  skills: DashboardSummarySkills;
  notes: DashboardSummaryNotes;
  recentActivity: DashboardSummaryRecentActivityItem[];
}