# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeApp API — skills, tasks, notes, translation, AI assistant
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: number;
  skillId: number;
  title: string;
  content: string;
  order: number;
  isCompleted: boolean;
  /** @nullable */
  duration?: string | null;
}