# lib/api-zod/src/generated/types/noteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeApp API — skills, tasks, notes, translation, AI assistant
 * OpenAPI spec version: 0.1.0
 */

export interface NoteUpdate {
  title?: string;
  content?: string;
  tag?: string;
}