# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeApp API — skills, tasks, notes, translation, AI assistant
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  id: number;
  name: string;
  description: string;
  category: string;
  icon: string;
  totalLessons: number;
  completedLessons: number;
}